+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Added
 
+- Kaocha watch can now add ignores from `.gitignore` and `.ignore`. To enable
+    this feature, set `:kaocha.watch/use-ignore-file` to true in your deps.edn.
+
 ## Fixed
 
 - Clearly alert the user that Clojure versions before 1.9 aren't supported, rather than